// src/api/ExceptionType.ts
export enum ExceptionType {
  Base = 'Base',
  ArithmeticError = 'ArithmeticError',
  BufferError = 'BufferError',
  LookupError = 'LookupError',
  AssertionError = 'AssertionError',
  AttributeError = 'AttributeError',
  EofError = 'EofError',
  FloatingPointError = 'FloatingPointError',
  GeneratorExit = 'GeneratorExit',
  ImportError = 'ImportError',
  ModuleNotFoundError = 'ModuleNotFoundError',
  IndexError = 'IndexError',
  KeyError = 'KeyError',
  KeyboardInterrupt = 'KeyboardInterrupt',
  MemoryError = 'MemoryError',
  NameError = 'NameError',
  NotImplementedError = 'NotImplemented',
  OsError = 'OsError',
  OverflowError = 'OverflowError',
  RecursionError = 'RecursionError',
  ReferenceError = 'ReferenceError',
  RuntimeError = 'RuntimeError',
  StopIteration = 'StopIteration',
  StopAsyncIteration = 'StopAsyncIteration',
  SyntaxError = 'SyntaxError',
  IndentationError = 'IndentationError',
  TabError = 'TabError',
  SystemError = 'SystemError',
  SystemExit = 'SystemExit',
  TypeError = 'TypeError',
  UnboundLocalError = 'UnboundLocalError',
  UnicodeError = 'UnicodeError',
  UnicodeEncodeError = 'UnicodeEncodeError',
  UnicodeDecodeError = 'UnicodeDecodeError',
  UnicodeTranslateError = 'UnicodeTranslateError',
  ValueError = 'ValueError',
  ZeroDivisionError = 'ZeroDivision',
  EnvironmentError = 'EnvironmentError',
  IoError = 'IoError',
  BlockingIoError = 'BlockingIoError',
  ChildProcessError = 'ChildProcessError',
  ConnectionError = 'ConnectionError',
  BrokenPipeError = 'BrokenPipeError',
  ConnectionAbortedError = 'ConnectionAborted',
  ConnectionRefusedError = 'ConnectionRefused',
  ConnectionResetError = 'ConnectionReset',
  FileExistsError = 'FileExists',
  FileNotFoundError = 'FileNotFound',
  InterruptedError = 'Interrupted',
  IsADirectoryError = 'IsADirectoryError',
  NotADirectoryError = 'NotADirectoryError',
  PermissionError = 'PermissionError',
  ProcessLookupError = 'ProcessLookupError',
  TimeoutError = 'TimeoutError',
  NotASequence = 'NotASequence',
  NotAFunction = 'NotAFunction',
  UnknownIdentifier = 'UnknownIdentifier',
  FunctionArgumentError = 'FunctionArgumentError',
  FunctionMissingArgument = 'FunctionMissingArgument',
  FunctionDuplicateArgumentError = 'FunctionDuplicateArgumentError',
  ExpectedReference = 'ExpectedReference',
  UnpackSourceIsNotSequence = 'UnpackSourceIsNotSequence',
  CannotUnpackToEmptyTuple = 'CannotUnpackToEmptyTuple',
  UnpackCountDoesntMatch = 'UnpackCountDoesntMatch',
  FunctionArgumentCountMismatch = 'FunctionArgumentCountMismatch',
  FunctionTooManyArguments = 'FunctionTooManyArguments',
  ImportAllowedOnlyOnModuleLevel = 'ImportAllowedOnlyOnModuleLevel',
  CannotBuildClassHierarchy = 'CannotBuildClassHierarchy',
  CannotDeriveFromMultipleException = 'CannotDeriveFromMultipleException',
  BreakOrContinueOutsideOfCycle = 'BreakOrContinueOutsideOfCycle',
  ResolutionOrder = 'ResolutionOrder',
  CannotReRaise = 'CannotReRaise',

  NonPublicExceptionType = 'NonPublicExceptionType', // not an exception, used for getExceptionType only
}